pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lexic-api"
version = "0.1.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
description = "Lexic API for lawyer platform"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn>=0.27.1",
    "sqlalchemy>=2.0.27",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.1",
    "pydantic[email]>=2.6.1",
    "pydantic-settings>=2.1.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]